Can click on a specific game now
Made link_path unique in the database for games table. Changed name on variable 'games_id' to 'game_id' to follow Laravel naming convention. Made it possible to click on a specific game and then get presented with the chosen game's information. Seperated the category bar from the master layout since it shouldn't be loaded on gamepage.blade.