I have just made alot of changes.
I have managed to create a Theme-function for my dropdown menu, so that the user is able to switch between the dark and light (default) mode.

My next target is to try and implement rgb, so that the user is able to select more colors instead of the ones which are defined in a default way.